+++ InvenTree/web/src/src/pages/Index/Profile/SettingsPanel.tsx
@@ -1,0 +1,10 @@
+import { Container, Text } from '@mantine/core';
+import { Trans } from '@lingui/macro';
+
+export function SettingsPanel() {
+    return (
+        <Container>
+            <Text><Trans>Settings</Trans></Text>
+        </Container>
+    );
+}
